# Conflicts:
#	core/src/main/java/uk/ac/york/sepr4/hud/HUD.java

+++ core/src/main/java/uk/ac/york/sepr4/object/building/BuildingManager.java
@@ -112,7 +112,7 @@ public class BuildingManager {
             Optional<Vector2> pos = getValidRandomSpawn(college, 250f);
             if(pos.isPresent()) {
                 NPCBoat boat = new NPCBuilder().generateRandomEnemy( pos.get(), college,
-                                              boss ? college.getBossDifficulty() : college.getEnemyDifficulty(), boss);
+                         boss ? college.getBossDifficulty().floatValue() : college.getEnemyDifficulty().floatValue(), boss);
                 return Optional.of(boat);
             }
         }
